Show off the new shrinkage
While I really don’t like the source appearance of this output, it does
greatly reduce filesize (which is our/my goal), so ship it.
Not ready to do it to all files yet, as I am still playing with
appearance and mehtods for styling - this technique is better suited to
heavily h/v paths.
-rw-rw-r-- 1 john 3587 2015-02-20 15:10 johndrinkwater.svg
-rw-rw-r-- 1 john 2568 2015-02-24 10:39 johndrinkwater.svg