StumbleUpon gets a logo
Know this is kinda against their guidelines, but all our other logos are
Brand colour + shape, with the presentation side of it making the true
logo.

It has been a while, huh? Desktop was out of action, OSX w/o Inkscape
and my usual tools meant there was a mild holiday.